requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
pypdf>=3.17.4
pdf2docx>=0.5.8
Pillow>=10.2.0
PyMuPDF>=1.23.8
weasyprint>=60.1
